jwt 토큰을 이용하여 모두의 장바구니 api에서 로그인 된 개인의 장바구니api로 수정

+++ routes/goods.js
@@ -1,12 +1,42 @@
 const express = require("express");
 const router = express.Router();
 const goodsSchema = require("../schemas/goods");
+const Cart = require("../schemas/cart.js");
+const authMiddleware = require("../middlewares/auth-middleware");
+
+//장바구니 조회 API
+router.get("/goods/cart", authMiddleware, async(req,res) => {
+  const { userId } = res.locals.user;
+  const carts = await Cart.find({userId : userId});
+  // [
+  //  {goodsId, quantity},
+  //  {goodsId, quantity},
+  // ];
+  const goodsIds = carts.map((cart) => {
+      return cart.goodsId;
+  })
+  // [2, 11, 19];
+
+  const goods = await goodsSchema.find({goodsId: goodsIds});
+  // Goods에 해당하는 모든 정보를 가지고 올건데,
+  // 만약 goodsIds 변수 안에 존재하는 값일 때에만 조회하라.
+
+  const results = carts.map((cart) => {
+      return {
+          "quantity": cart.quantity,
+          "goods": goods.find((item) => item.goodsId === cart.goodsId),
+      }
+  })
+
+  res.json({
+      "carts": results,
+  })
+
+});
 
 //상품 목록 조회 API
 router.get("/goods", async(req, res) => {
   const { category } = req.query;
-  console.log(category);
-
 // category가 존재할 때에는 category에 맞게 조회하고, 아니면 전체를 조회한다.
   const goods = await goodsSchema.find(category ? {category}: {}) 
   .sort("-date") // 내림 차순으로 정렬
@@ -44,15 +74,14 @@ router.get("/goods/:goodsId", async(req, res) => {
   res.status(200).json({goods: result});
 });
 
-
-
-const Cart = require("../schemas/cart.js");
-
-router.post("/goods/:goodsId/cart", async (req, res) => {
+//장바구니 등록 API
+router.post("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  // 사용자 정보(userId)를 가지고 장바구니를 조회한다
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
     return res.status(400).json({
       success: false,
@@ -60,36 +89,43 @@ router.post("/goods/:goodsId/cart", async (req, res) => {
     })
   }
 
-  await Cart.create({ goodsId, quantity });
+  //해당하는 사용자의 정보(userId)를 가지고, 장바구니에 상품을 등록한다.
+  await Cart.create({ userId, goodsId, quantity });
 
   res.json({ result: "success" });
 })
 
-router.put("/goods/:goodsId/cart", async (req, res) => {
+//장바구니 수정 API
+router.put("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
     await Cart.updateOne(
-      { goodsId: goodsId },
+      { userId, goodsId: goodsId },
       { $set: { quantity: quantity } }
     )
   }
   res.status(200).json({ success: true });
 })
 
-router.delete("/goods/:goodsId/cart", async (req, res) => {
+//장바구니 삭제 API
+router.delete("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
-    await Cart.deleteOne({ goodsId });
+    await Cart.deleteOne({ userId, goodsId });
   }
 
   res.json({ result: "success" });
 })
 
+
+//goods 생성 API
 router.post("/goods", async (req, res) => {
   const { goodsId, name, thumbnailUrl, category, price } = req.body;
